+++ src/commons/index.ts
@@ -1,0 +1,3 @@
+export * from './entity'
+export * from './service.commons'
+export * from './controller.commons'
