+++ src/modules/admin/admin.controller.ts
@@ -1,9 +1,8 @@
 import { Controller } from '@nestjs/common';
 
 import {AdminService} from "./admin.service";
-import {BaseController} from "../../commons/controller.commons";
-import {BaseService} from "../../commons/service.commoms";
-import {Admin} from "./entities/admin.entity";
+import {BaseController, BaseService} from "../../commons";
+import {Admin} from "./entities";
 
 @Controller('admin')
 export class AdminController extends BaseController<Admin>{
