+++ src/modules/admin/admin.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import {TypeOrmModule} from "@nestjs/typeorm";
-import {Admin} from "./entities/admin.entity";
+import {Admin} from "./entities";
 import {AdminController} from "./admin.controller";
 import {AdminService} from "./admin.service";
