+++ src/modules/admin/index.ts
@@ -1,0 +1,4 @@
+export * from './entities'
+export * from './admin.service'
+export * from './admin.module'
+export * from './admin.controller'
