--- src/modules/assignment/assignment.controller.spec.ts
@@ -1,20 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AssignmentController } from './assignment.controller';
-import { AssignmentService } from './assignment.service';
-
-describe('AssignmentController', () => {
-  let controller: AssignmentController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [AssignmentController],
-      providers: [AssignmentService],
-    }).compile();
-
-    controller = module.get<AssignmentController>(AssignmentController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
