--- src/modules/assignment/assignment.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { AssignmentService } from './assignment.service';
-
-describe('AssignmentService', () => {
-  let service: AssignmentService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [AssignmentService],
-    }).compile();
-
-    service = module.get<AssignmentService>(AssignmentService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
