+++ src/modules/assignment/entities/assignment.entity.ts
@@ -1,1 +1,6 @@
-export class Assignment {}
+import {BaseEntity} from "../../../commons/";
+
+export class Assignment extends BaseEntity{
+  id: number;
+
+}
