+++ src/modules/assignment/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './assignment.entity'
