+++ src/modules/assignment/index.ts
@@ -1,0 +1,4 @@
+export * from './assignment.controller'
+export * from './assignment.module'
+export * from './assignment.service'
+export * from './entities'
