+++ src/modules/index.ts
@@ -1,0 +1,4 @@
+export * from './jtp'
+export * from './admin'
+export * from './student'
+export * from './assignment'
