+++ src/modules/jtp/entities/index.ts
@@ -1,0 +1,1 @@
+export * from './jtp.entity'
