+++ src/modules/jtp/index.ts
@@ -1,0 +1,5 @@
+export * from './jtp.module'
+export * from './jtp.controller'
+export * from './jtp.service'
+export * from './jtp.module'
+export * from './entities'
