+++ src/modules/jtp/jtp.controller.ts
@@ -1,8 +1,7 @@
 import {Controller} from '@nestjs/common';
 import {JtpService} from './jtp.service';
-import {BaseController} from "../../commons/controller.commons";
-import {Jtp} from "./entities/jtp.entity";
-import {BaseService} from "../../commons/service.commoms";
+import {BaseController, BaseService} from "../../commons";
+import {Jtp} from "./entities";
 
 @Controller('jtp')
 export class JtpController extends BaseController<Jtp> {
