+++ src/modules/jtp/jtp.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { JtpService } from './jtp.service';
 import { JtpController } from './jtp.controller';
 import {TypeOrmModule} from "@nestjs/typeorm";
-import {Jtp} from "./entities/jtp.entity";
+import {Jtp} from "./entities";
 
 @Module({
   imports:[TypeOrmModule.forFeature([Jtp])],
