+++ src/modules/student/index.ts
@@ -1,0 +1,4 @@
+export * from './student.service'
+export * from './student.controller'
+export * from './student.service'
+export * from './entities'
