+++ src/modules/student/student.controller.ts
@@ -1,12 +1,8 @@
-import { Controller, Get, Post, Body, Patch, Param, Delete } from '@nestjs/common';
+import { Controller} from '@nestjs/common';
 import { StudentService } from './student.service';
-import { CreateStudentDto } from './dto/create-student.dto';
-import { UpdateStudentDto } from './dto/update-student.dto';
-import {BaseController} from "../../commons/controller.commons";
-import {Student} from "./entities/student.entity";
-import { BaseService } from 'src/commons/service.commoms';
-import {AdminService} from "../admin/admin.service";
-import {Admin} from "../admin/entities/admin.entity";
+import {BaseController} from "../../commons";
+import {Student} from "./entities";
+import { BaseService } from 'src/commons/service.commons';
 
 @Controller('students')
 export class StudentController extends BaseController<Student> {
