+++ src/modules/student/student.module.ts
@@ -1,8 +1,8 @@
-import { Module } from '@nestjs/common';
-import { StudentService } from './student.service';
-import { StudentController } from './student.controller';
+import {Module} from '@nestjs/common';
+import {StudentService} from './student.service';
+import {StudentController} from './student.controller';
 import {TypeOrmModule} from "@nestjs/typeorm";
-import {Student} from "./entities/student.entity";
+import {Student} from "./entities";
 
 @Module({
   imports: [TypeOrmModule.forFeature([Student])],
