--- src/modules/student/student.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { StudentService } from './student.service';
-
-describe('StudentService', () => {
-  let service: StudentService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [StudentService],
-    }).compile();
-
-    service = module.get<StudentService>(StudentService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
